+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Change some function arguments and return for faster processing.
+
 ## [1.0.0-alpha.7] - 2022-12-25
 
 ### Changed